[python] #28 boardInterpreter
