+++ playground/house.py
@@ -76,7 +76,8 @@ class House:
                 break
 
             if render_every:
-                self.board.render_figure()
+                #self.board.render_figure()
+                self.board.draw_board(self.board.log)
                 time.sleep(pause)
 
 
@@ -84,4 +85,4 @@ if __name__ == "__main__":
     coin_toss = True if random() > 0.5 else False
     print("1p first: ", coin_toss)
     house = House(first_player_first_move=coin_toss)
-    house.simulate(render_every=False, pause=0.5)
+    house.simulate(render_every=True, pause=0.5)
